feat: implement plan and transaction retrieval methods in repositories and handlers

+++ src/payloads/transaction_payloads.rs
@@ -1,5 +1,6 @@
 use chrono::{DateTime, Utc};
 use serde::Serialize;
+use sqlx::types::BigDecimal;
 use ts_rs::TS;
 
 use super::enums;
@@ -10,7 +11,7 @@ pub struct UserTransaction {
     pub id: u32,
     pub user_id: u32,
     pub plan_id: u32,
-    pub amount: f64,
+    pub amount: BigDecimal,
     pub status: enums::UserTransactionStatus,
     pub stripe_payment_intent_id: Option<String>,
     pub created_at: DateTime<Utc>,
